FUT1-1617 DeviceUseStatement add context(CarePlan) When a Device is used by a Patient, we need to know the context (i.e. the CarePlan).
Therefore DeviceUseStatement must be extended 1..1 with field 'Reference(CarePlan): context'.

If a Device is used by the same Patient for more than 1 CarePlan, then more than one DeviceUseStatement will be created and have to have same CarePlan as context.

Device.patient must be profiled to a 0..0 relation since we will not use it.